Even points (half j)

+++ src/LKrav.jl
@@ -4,17 +4,17 @@ export  ϕd, ϕu, LKd, LKu, LK
 
 "`Φd(q₁, q₂, n, m, j)` shows the rotations of the Kravchuk function modes at (n, |m| ≤ n) for n ∈ [0, 2j]. It needs to be evaluted at (q₁, q₂) ∈ [0, 2j + 1]."
 function Φd(q₁, q₂, n, m, j)
-    v₀ = zeros(2 * j + 1)
+    v₀ = zeros(Integer(2 * j + 1)) # (Sep 9, 2021) Added Integer() to meet the half-spin feature in j
     for n₁ in 0:n
-        v₀[n₁ + 1] = Φ2(n₁, n - n₁, q₁, q₂, j) * dd(n, j)[n₁ + 1, Integer((m + n)/2) + 1]
+        v₀[Integer(n₁ + 1)] = Φ2(n₁, n - n₁, q₁, q₂, j) * dd(n, j)[n₁ + 1, Integer((m + n)/2) + 1]
     end
     return sum(v₀)
 end
 
 "`Φu(q₁, q₂, n, m, j)` shows the rotations of the Kravchuk function modes at (n, |m| ≤ 4j - n) for n ∈ [4j, 2j]. It needs to be evaluted at (q₁, q₂) ∈ [0, 2j + 1]."
 function Φu(q₁, q₂, n, m, j)
-    v₁ = zeros(2 * j + 1)
-    for n₁ in 0:(4* j - n)
+    v₁ = zeros(Integer(2 * j + 1))
+    for n₁ in 0:Integer((4* j - n))
         v₁[n₁ + 1] = Φ2(2 * j - n₁, n + n₁ - 2 * j, q₁, q₂, j) * du(n, j)[n₁ + 1, Integer((m + 4 * j - n)/2) + 1]
     end
     return sum(v₁)
@@ -24,7 +24,7 @@ end
 
 "`LKd(q₁, q₂, n, m, j)` shows the Laguerre-Kravchuk modes at (n, |m| ≤ n) for n ∈ [0, 2j]. It needs to be evaluted at (q₁, q₂) ∈ [0, 2j + 1]."
 function LKd(q₁, q₂, n, m, j)
-    vlkd = zeros(ComplexF64, 2 * j + 1)
+    vlkd = zeros(ComplexF64, Integer(2 * j + 1))
     for n₁ in 0:n
         vlkd[n₁ + 1] = (-1.0im)^(n - n₁) * Φ2(n₁, n - n₁, q₁, q₂, j) * dd(n, j)[n₁ + 1, Integer((m + n)/2) + 1]
     end
@@ -33,8 +33,8 @@ end
 
 "`LKu(q₁, q₂, n, m, j)` shows the Laguerre-Kravchuk modes at (n, |m| ≤ 4j - n) for n ∈ [4j, 2j]. It needs to be evaluted at (q₁, q₂) ∈ [0, 2j + 1]."
 function LKu(q₁, q₂, n, m, j)
-    vlku = zeros(ComplexF64, 2 * j + 1)
-    for n₁ in 0:(4* j - n)
+    vlku = zeros(ComplexF64, Integer(2 * j + 1))
+    for n₁ in 0:Integer(4* j - n)
         vlku[n₁ + 1] = (-1.0im)^(n - n₁) * Φ2(2 * j - n₁, n + n₁ - 2 * j, q₁, q₂, j) * du(n, j)[n₁ + 1, Integer((m + 4 * j - n)/2) + 1]
     end
     return (-1.0)^((0.5) * (abs(m) - m)) * sum(vlku)
